backend/requirements.txt
asgiref==3.5.2
boto3==1.26.118
botocore==1.29.163
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
dj-database-url==1.3.0
Django==4.1.2
django-cors-headers==3.13.0
django-decouple==2.1
django-filter==23.2
django-rest-swagger==2.2.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
msrest==0.7.1
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==21.3
Pillow==9.2.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
